Add Busqueda por id prices
se agrega la busqueda por id de precio

+++ comercio-electronico/src/main/java/com/comercio/electronico/model/Prices.java
@@ -18,7 +18,7 @@ public class Prices {
 	
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private Long priceList;
+	private Integer priceList;
 	private String name;
 	private Date startDate;
 	private Date endDate;
